+++ CPDShell/Core/algorithms/KNNCPD/knn_heap.py
@@ -9,7 +9,7 @@ __license__ = "SPDX-License-Identifier: MIT"
 import heapq
 import typing as tp
 
-from .observation import Observation, Neighbour
+from .abstracts.observation import Observation, Neighbour
 
 
 class NNHeap:
@@ -51,7 +51,7 @@ class NNHeap:
         :param observation: observation to test.
         """
         def predicate(x: Neighbour) -> bool:
-            return x.observation is observation
+            return x.observation.value is observation.value
 
         return any(predicate(i) for i in self._heap)
 
@@ -68,20 +68,6 @@ class NNHeap:
         neighbour = Neighbour(neg_distance, observation)
 
         if len(self._heap) == self._size and neighbour.distance >= self._heap[0].distance:
-            old_neighbour = heapq.heapreplace(self._heap, neighbour)
-            self.__add_auxiliary(old_neighbour)
-        elif len(self._heap) == self._size and neighbour.distance < self._heap[0].distance:
-            self.__add_auxiliary(neighbour)
-        else:
+            heapq.heapreplace(self._heap, neighbour)
+        elif len(self._heap) < self._size:
             heapq.heappush(self._heap, neighbour)
-
-    def __add_auxiliary(self, neighbour: Neighbour) -> None:
-        """
-        Adds neighbour to auxiliary heap.
-
-        :param neighbour: Neighbour to add.
-        """
-        if len(self._auxiliary_heap) == self._size and neighbour.distance >= self._auxiliary_heap[0].distance:
-            heapq.heapreplace(self._auxiliary_heap, neighbour)
-        elif len(self._auxiliary_heap) < self._size:
-            heapq.heappush(self._auxiliary_heap, neighbour)
